Address MSVC build issues and warnings

+++ ohm/NdtMap.h
@@ -60,9 +60,9 @@ namespace ohm
     /// Set the range sensor noise estimate. For example, the range noise for a lidar sensor.
     ///
     /// @param noise_range The sensor noise range. Must be greater than zero.
-    void setSensorNoise(double noise_range);
+    void setSensorNoise(float noise_range);
     /// Read the range sensor noise estimate.
-    double sensorNoise() const;
+    float sensorNoise() const;
 
     /// Set the number of samples required in a voxel before using the NDT algorithm for @c integateMiss() adjustments.
     void setNdtSampleThreshold(unsigned sample_count);
